1: Refactored calculators by removing code duplication.
