+++ src/EveryFan.Recruitment/EveryFan.Recruitment.csproj
@@ -41,6 +41,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="PayingPosition.cs" />
+    <Compile Include="PayoutCalculators\PayoutCalculator.cs" />
     <Compile Include="PayoutEngine.cs" />
     <Compile Include="PayoutCalculators\FiftyFiftyPayoutCalculator.cs" />
     <Compile Include="PayoutCalculators\IPayoutCalculator.cs" />
