+++ src/EveryFan.Recruitment/PayoutCalculators/WinnerTakesAllPayoutCalculator.cs
@@ -8,26 +8,11 @@ namespace EveryFan.Recruitment.PayoutCalculators
     /// Winner takes all payout calculator, the winner recieves the entire prize pool. In the event of a tie for the winning position the
     /// prize pool is split equally between the tied players.
     /// </summary>
-    public class WinnerTakesAllPayoutCalculator : IPayoutCalculator
+    public class WinnerTakesAllPayoutCalculator : PayoutCalculator
     {
-        private IReadOnlyList<PayingPosition> GetPayingPositions(Tournament tournament)
+        protected override IReadOnlyList<PayingPosition> GetPayingPositions(Tournament tournament)
         {
             throw new NotImplementedException();
-        }
-
-        public IReadOnlyList<TournamentPayout> Calculate(Tournament tournament)
-        {
-            IReadOnlyList<PayingPosition> payingPositions = this.GetPayingPositions(tournament);
-            IReadOnlyList<TournamentEntry> orderedEntries = tournament.Entries.OrderByDescending(p => p.Chips).ToList();
-
-            List<TournamentPayout> payouts = new List<TournamentPayout>();
-            payouts.AddRange(payingPositions.Select((p, i) => new TournamentPayout()
-            {
-                Payout = p.Payout,
-                UserId = orderedEntries[i].UserId
-            }));
-
-            return payouts;
         }
     }
 }
